indoor_style_recognition/requirements.txt
pandas>=1.3.0
openpyxl>=3.1.0
pillow>=8.3.0
torch>=1.9.0
torchvision>=0.10.0
timm>=0.9.16
numpy>=1.20.0
scikit-learn>=0.24.0
tqdm>=4.61.0
matplotlib>=3.4.0
requests>=2.26.0
tensorboard>=2.6.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.0
